+++ packages/KetcherSketcher/src/ketcher.tsx
@@ -37,11 +37,11 @@ export class KetcherSketcher extends grok.chem.SketcherBase {
       },
       onInit: (ketcher: Ketcher) => {
         this._sketcher = ketcher;
-        grok.dapi.userDataStorage.getValue(KETCHER_OPTIONS, KETCHER_USER_STORAGE, true).then((opts: string) => {
-          if (opts) {
-            this._sketcher?.editor.setOptions(opts);
-          }
-        });
+        // grok.dapi.userDataStorage.getValue(KETCHER_OPTIONS, KETCHER_USER_STORAGE, true).then((opts: string) => {
+        //   if (opts) {
+        //     this._sketcher?.editor.setOptions(opts);
+        //   }
+        // });
         this.setMoleculeFromHost();
         //@ts-ignore
         window[KETCHER_WINDOW_OBJECT] = ketcher;
@@ -174,7 +174,7 @@ export class KetcherSketcher extends grok.chem.SketcherBase {
   }
 
   detach() {
-    grok.dapi.userDataStorage.postValue(KETCHER_OPTIONS, KETCHER_USER_STORAGE, JSON.stringify(this._sketcher?.editor.options()), true);
+   // grok.dapi.userDataStorage.postValue(KETCHER_OPTIONS, KETCHER_USER_STORAGE, JSON.stringify(this._sketcher?.editor.options()), true);
     super.detach();
   }