+++ packages/Peptides/src/viewers/logo-summary.ts
@@ -569,6 +569,7 @@ export class LogoSummaryTable extends DG.JsViewer implements ILogoSummaryTable {
     grid.props.rowHeight = 55;
 
     const webLogoCache = new DG.LruCache<number, DG.Viewer & IWebLogoViewer>();
+    // @ts-ignore TODO: fix after api update
     const distCache = new DG.LruCache<number, DG.Viewer<DG.IHistogramSettings>>();
     const maxSequenceLen = this.positionColumns.length;
     const webLogoGridCol = grid.columns.byName(C.LST_COLUMN_NAMES.WEB_LOGO)!;
@@ -628,6 +629,7 @@ export class LogoSummaryTable extends DG.JsViewer implements ILogoSummaryTable {
               });
             webLogoCache.set(currentRowIdx, viewer);
           }
+          viewer.root.style.height = `${height}px`;
           gridCell.element = viewer.root;
           gridCellArgs.preventDefault();
         } else if (gridCell.tableColumn?.name === C.LST_COLUMN_NAMES.DISTRIBUTION) {