+++ packages/Peptides/src/widgets/distribution.ts
@@ -91,7 +91,9 @@ export function getDistributionWidget(table: DG.DataFrame, options: Distribution
  * @param isTooltip - Is histogram for tooltip
  * @return - Histogram viewer
  */
-export function getActivityDistribution(table: DG.DataFrame, isTooltip: boolean = false,
+export function getActivityDistribution(
+  table: DG.DataFrame, isTooltip: boolean = false,
+  // @ts-ignore TODO: fix after api update
 ): DG.Viewer<DG.IHistogramSettings> {
   const hist = table.plot.histogram({
     filteringEnabled: false,
@@ -103,6 +105,7 @@ export function getActivityDistribution(table: DG.DataFrame, isTooltip: boolean 
     showRangeSlider: false,
     showBinSelector: false,
     backColor: isTooltip ? '#fdffe5' : '#fffff',
+    // @ts-ignore TODO: fix after api update
   }) as DG.Viewer<DG.IHistogramSettings>;
   hist.root.style.width = 'auto';
   return hist;