+++ packages/UsageAnalysis/src/tests/app.ts
@@ -1,15 +1,16 @@
 import * as DG from 'datagrok-api/dg';
 import * as grok from 'datagrok-api/grok';
 
-import {before, after, category, test, expect, awaitCheck} from '@datagrok-libraries/utils/src/test';
-import {ViewHandler} from '../view-handler';
+import { before, after, category, test, expect, awaitCheck } from '@datagrok-libraries/utils/src/test';
+import { ViewHandler } from '../view-handler';
 
 category('App', () => {
-  const tabs = ['Overview', 'Packages', 'Functions', 'Events', 'Log', 'Tests', 'Reports'];
-  const num = [4, 4, 5, 4, 3, 5, 7];
-  const handler = new ViewHandler();
+  const tabs = ['Overview', 'Packages', 'Functions', 'Events', 'Log', 'Tests'];
+  const num = [4, 4, 5, 4, 3, 5];
+  let handler = new ViewHandler();
 
   before(async () => {
+    handler = new ViewHandler();
     if (grok.shell.sidebar.panes.every((p) => p.name != ViewHandler.UA_NAME)) {
       await handler.init();
       grok.shell.addView(handler.view);
@@ -21,13 +22,12 @@ category('App', () => {
     expect(grok.shell.v.name === 'Usage Analysis', true, 'cannot find Usage Analysis view');
   });
 
-  for (let i = 0; i < 7; i++) {
+  for (let i = 0; i < 6; i++) {
     test(tabs[i], async () => {
       handler.changeTab(tabs[i]);
       const view = grok.shell.v.root;
-      let err = null;
-      let s = ['Log', 'Errors'].includes(tabs[i]) ? '.grok-wait + .d4-grid canvas' : 'canvas';
-      s = tabs[i] === 'Reports' ? '.d4-grid' : s;
+      let err: any = undefined;
+      let s = ['Log'].includes(tabs[i]) ? '.grok-wait + .d4-grid canvas' : 'canvas';
       try {
         await awaitCheck(() => view.querySelectorAll(s).length === num[i], '', 10000);
       } catch (e) {
@@ -45,4 +45,4 @@ category('App', () => {
     DG.Balloon.closeAll();
     grok.shell.windows.showContextPanel = true;
   });
-}, {clear: false, timeout: 60000});
+}, { clear: false, timeout: 60000 });