+++ packages/EnamineStore/CHANGELOG.md
@@ -1,5 +1,9 @@
 # EnamineStore changelog
 
+## 1.1.0 (2024-05-24)
+
+Updated swagger file according to new Enamine Store API
+
 ## 1.0.2 (2023-07-24)
 
 This release focuses on improving package usability.